Updating Queues DB for Imported Batch Logs
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C25490

+++ TOneV2/Code/TOneV2/SQL.TOneLogging/integration/Tables/DataSourceLog.sql
@@ -3,9 +3,11 @@
     [DataSourceId]    INT            NOT NULL,
     [Severity]        INT            NOT NULL,
     [Message]         VARCHAR (1000) NULL,
-    [ImportedBatchId] INT            NULL,
+    [ImportedBatchId] BIGINT         NULL,
     [LogEntryTime]    DATETIME       NULL,
     CONSTRAINT [PK_DataSourceLog] PRIMARY KEY CLUSTERED ([ID] ASC),
     CONSTRAINT [FK_DataSourceLog_DataSourceImportedBatch] FOREIGN KEY ([ImportedBatchId]) REFERENCES [integration].[DataSourceImportedBatch] ([ID])
 );
+
+
